:truck: separate types & index for export on button
