+++ packages/react/src/components/Spacing/spacing.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
-import Button from '../Button';
+import { Button } from '../Button';
 import { FlexBox } from '../FlexBox';
 import Paper from '../Paper';
 import { Text } from '../Text';
